requirements.txt
# Основные зависимости для работы с API
spotipy==2.23.0
yandex-music==2.2.0
python-dotenv==1.0.0

# Утилита для получения токена Яндекс.Музыка (не требуется в рантайме)
# yandex-music-token

# HTTP клиент для дополнительных запросов
requests==2.31.0

# Логирование и обработка ошибок
structlog==23.2.0

# Валидация данных
pydantic==2.5.0

# Асинхронность (для будущих улучшений)
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Web server for HTTP interface
Flask==2.3.3

# Разработка
black==23.12.1
flake8==6.1.0
mypy==1.8.0
